Project/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
bcrypt==3.2.0
blinker==1.4
cachelib==0.1.1
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cssselect==1.1.0
cssutils==1.0.2
distlib==0.3.1
Django==3.1.7
dnspython==2.1.0
email-to==0.1.0
email-validator==1.1.2
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
flask-blueprint==1.3.0
Flask-Login==0.5.0
Flask-Login-Multi==0.1.2
Flask-Mail==0.9.1
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
Flask-Session==0.3.2
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
image==1.5.33
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
keyboard==0.13.5
lazy-object-proxy==1.4.3
lxml==4.6.2
Markdown==3.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
mongoengine==0.23.0
numpy==1.20.1
onetimepass==1.0.1
opencv-contrib-python==4.5.1.48
pbr==5.5.1
Pillow==8.1.0
premailer==3.7.0
pycparser==2.20
pylint==2.6.0
pytz==2021.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.22
SQLAlchemy-ImageAttach==1.1.0
SQLAlchemy-Utils==0.36.8
sqlparse==0.4.1
stevedore==3.3.0
toml==0.10.2
urllib3==1.26.2
virtualenv==20.4.2
virtualenv-clone==0.5.4
Wand==0.6.5
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
